+++ knotx-adapter/knotx-adapter-service-http/src/main/java/io/knotx/adapter/service/http/impl/HttpServiceAdapterProxyImpl.java
@@ -20,20 +20,19 @@ import io.knotx.adapter.common.http.HttpAdapterConfiguration;
 import io.knotx.adapter.common.http.HttpClientFacade;
 import io.knotx.dataobjects.AdapterRequest;
 import io.knotx.dataobjects.AdapterResponse;
+import io.reactivex.Single;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.client.WebClientOptions;
-import io.vertx.rxjava.core.Vertx;
-import io.vertx.rxjava.ext.web.client.WebClient;
-import rx.Single;
+import io.vertx.reactivex.core.Vertx;
+import io.vertx.reactivex.ext.web.client.WebClient;
 
 public class HttpServiceAdapterProxyImpl extends AbstractAdapterProxy {
 
   private HttpClientFacade httpClientFacade;
 
   public HttpServiceAdapterProxyImpl(Vertx vertx, HttpAdapterConfiguration configuration) {
-    this.httpClientFacade = new HttpClientFacade(getWebClient(vertx, configuration),
-        configuration.getServices());
+    this.httpClientFacade = new HttpClientFacade(getWebClient(vertx, configuration), configuration);
   }
 
   @Override