+++ knotx-core/src/main/java/io/knotx/dataobjects/KnotContext.java
@@ -18,11 +18,11 @@ package io.knotx.dataobjects;
 import com.google.common.base.Objects;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
+import io.reactivex.Single;
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
 import java.util.List;
 import java.util.Set;
-import rx.Single;
 
 @DataObject(generateConverter = true)
 public class KnotContext {