+++ knotx-example/knotx-example-action-adapter-http/src/main/java/io/knotx/adapter/action/http/HttpActionAdapterVerticle.java
@@ -49,7 +49,7 @@ public class HttpActionAdapterVerticle extends AbstractVerticle {
     //register the service proxy on event bus
     consumer = ProxyHelper
         .registerService(AdapterProxy.class, vertx,
-            new HttpActionAdapterProxyImpl(new io.vertx.rxjava.core.Vertx(vertx), configuration),
+            new HttpActionAdapterProxyImpl(new io.vertx.reactivex.core.Vertx(vertx), configuration),
             configuration.getAddress());
   }