+++ knotx-example/knotx-example-action-adapter-http/src/test/java/io/knotx/adapter/service/http/HttpClientFacadeTest.java
@@ -17,6 +17,7 @@ package io.knotx.adapter.service.http;
 
 import com.google.common.collect.Lists;
 import io.knotx.adapter.common.exception.UnsupportedServiceException;
+import io.knotx.adapter.common.http.HttpAdapterConfiguration;
 import io.knotx.adapter.common.http.HttpClientFacade;
 import io.knotx.adapter.common.http.ServiceMetadata;
 import io.knotx.dataobjects.AdapterRequest;
@@ -27,15 +28,16 @@ import io.knotx.junit.rule.Logback;
 import io.knotx.junit.rule.TestVertxDeployer;
 import io.knotx.junit.util.FileReader;
 import io.netty.handler.codec.http.HttpResponseStatus;
+import io.reactivex.Single;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.RunTestOnContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
-import io.vertx.rxjava.core.MultiMap;
-import io.vertx.rxjava.core.Vertx;
-import io.vertx.rxjava.ext.web.client.WebClient;
+import io.vertx.reactivex.core.MultiMap;
+import io.vertx.reactivex.core.Vertx;
+import io.vertx.reactivex.ext.web.client.WebClient;
 import java.util.Collections;
 import java.util.List;
 import java.util.regex.Pattern;
@@ -46,7 +48,6 @@ import org.junit.runner.RunWith;
 import org.mockito.Matchers;
 import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
-import rx.Single;
 
 @RunWith(VertxUnitRunner.class)
 public class HttpClientFacadeTest {
@@ -79,7 +80,8 @@ public class HttpClientFacadeTest {
     Async async = context.async();
     // given
     final WebClient mockedWebClient = PowerMockito.spy(webClient());
-    HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient, getServiceConfigurations());
+    HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient,
+        getConfiguration());
     final JsonObject expectedResponse = new JsonObject(FileReader.readText("first-response.json"));
 
     // when
@@ -88,15 +90,16 @@ public class HttpClientFacadeTest {
             HttpMethod.POST);
 
     // then
-    result.subscribe(
-        response -> {
+    result
+        .doOnSuccess(response -> {
           context.assertEquals(HttpResponseStatus.OK.code(), response.getStatusCode());
           context.assertEquals(expectedResponse, response.getBody().toJsonObject());
           Mockito.verify(mockedWebClient, Mockito.times(1))
               .request(HttpMethod.POST, PORT, DOMAIN, REQUEST_PATH);
-          async.complete();
-        },
-        error -> context.fail(error.getMessage()));
+        })
+        .subscribe(
+            response -> async.complete(),
+            error -> context.fail(error.getMessage()));
   }
 
   @Test
@@ -106,7 +109,8 @@ public class HttpClientFacadeTest {
     Async async = context.async();
     // given
     final WebClient mockedWebClient = PowerMockito.spy(webClient());
-    HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient, getServiceConfigurations());
+    HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient,
+        getConfiguration());
     final JsonObject expectedResponse = new JsonObject(FileReader.readText("first-response.json"));
 
     //FIXME - params to request
@@ -120,15 +124,16 @@ public class HttpClientFacadeTest {
                 HttpMethod.POST);
 
     // then
-    result.subscribe(
-        response -> {
+    result
+        .doOnSuccess(response -> {
           context.assertEquals(HttpResponseStatus.OK.code(), response.getStatusCode());
           context.assertEquals(expectedResponse, response.getBody().toJsonObject());
           Mockito.verify(mockedWebClient, Mockito.times(1))
               .request(HttpMethod.POST, PORT, DOMAIN, REQUEST_PATH);
-          async.complete();
-        },
-        error -> context.fail(error.getMessage()));
+        })
+        .subscribe(
+            response -> async.complete(),
+            error -> context.fail(error.getMessage()));
   }
 
   @Test
@@ -139,25 +144,24 @@ public class HttpClientFacadeTest {
     // given
     final WebClient mockedWebClient = PowerMockito.spy(webClient());
     HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient,
-        getServiceConfigurations());
+        getConfiguration());
 
     // when
     Single<ClientResponse> result = clientFacade.process(new AdapterRequest()
         .setParams(new JsonObject()).setRequest(clientRequest), HttpMethod.POST);
 
     // then
-    result.subscribe(
-        response -> context.fail("Error should occur!"),
-        error -> {
-          {
-            context
-                .assertEquals("Parameter `path` was not defined in `params`!", error.getMessage());
-            Mockito.verify(mockedWebClient, Mockito.times(0))
-                .request(Matchers.any(), Matchers.anyInt(), Matchers.anyString(),
-                    Matchers.anyString());
-            async.complete();
-          }
-        });
+    result
+        .doOnError(error -> {
+          context
+              .assertEquals("Parameter `path` was not defined in `params`!", error.getMessage());
+          Mockito.verify(mockedWebClient, Mockito.times(0))
+              .request(Matchers.any(), Matchers.anyInt(), Matchers.anyString(),
+                  Matchers.anyString());
+        })
+        .subscribe(
+            response -> context.fail("Error should occur!"),
+            error -> async.complete());
   }
 
   @Test
@@ -168,7 +172,7 @@ public class HttpClientFacadeTest {
     // given
     final WebClient mockedWebClient = PowerMockito.spy(webClient());
     HttpClientFacade clientFacade = new HttpClientFacade(mockedWebClient,
-        getServiceConfigurations());
+        getConfiguration());
 
     // when
     Single<ClientResponse> result =
@@ -177,24 +181,28 @@ public class HttpClientFacadeTest {
             HttpMethod.POST);
 
     // then
-    result.subscribe(
-        response -> context.fail("Error should occur!"),
-        error -> {
-          {
-            context.assertEquals(UnsupportedServiceException.class, error.getClass());
-            Mockito.verify(mockedWebClient, Mockito.times(0))
-                .request(Matchers.any(), Matchers.anyInt(), Matchers.anyString(),
-                    Matchers.anyString());
-            async.complete();
-          }
-        });
+    result
+        .doOnError(error -> {
+          context.assertEquals(UnsupportedServiceException.class, error.getClass());
+          Mockito.verify(mockedWebClient, Mockito.times(0))
+              .request(Matchers.any(), Matchers.anyInt(), Matchers.anyString(),
+                  Matchers.anyString());
+        })
+        .subscribe(
+            response -> context.fail("Error should occur!"),
+            error -> async.complete());
   }
 
   private WebClient webClient() {
     return WebClient.create(Vertx.newInstance(vertx.vertx()));
   }
+
   private AdapterRequest payloadMessage(JsonObject params, ClientRequest request) {
     return new AdapterRequest().setParams(params).setRequest(request);
+  }
+
+  private HttpAdapterConfiguration getConfiguration() {
+    return new HttpAdapterConfiguration().setServices(getServiceConfigurations());
   }
 
   private List<ServiceMetadata> getServiceConfigurations() {