+++ knotx-junit/src/main/java/io/knotx/junit/rule/TestVertxDeployer.java
@@ -20,14 +20,13 @@ import io.knotx.launcher.KnotxStarterVerticle;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.junit.RunTestOnContext;
-import io.vertx.rxjava.core.Vertx;
-import org.junit.rules.TestRule;
-import org.junit.runner.Description;
-import org.junit.runners.model.Statement;
-
+import io.vertx.reactivex.core.Vertx;
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+import org.junit.rules.TestRule;
+import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
 
 public class TestVertxDeployer implements TestRule {