+++ knotx-server/src/main/java/io/knotx/server/configuration/KnotxServerConfiguration.java
@@ -15,9 +15,10 @@
  */
 package io.knotx.server.configuration;
 
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.web.handler.BodyHandler;
 import java.util.Set;
 import java.util.stream.Collectors;
-import io.vertx.core.json.JsonObject;
 
 public class KnotxServerConfiguration {
 
@@ -29,11 +30,17 @@ public class KnotxServerConfiguration {
 
   private KnotxFlowConfiguration customFlow;
 
+  private String fileUploadDirectory;
+
   private JsonObject serverOptions;
 
+  private JsonObject customResponseHeader;
 
   public KnotxServerConfiguration(JsonObject config) {
     displayExceptionDetails = config.getBoolean("displayExceptionDetails", false);
+
+    customResponseHeader = config
+        .getJsonObject("customResponseHeader", new JsonObject());
 
     allowedResponseHeaders = config.getJsonArray("allowedResponseHeaders").stream()
         .map(item -> ((String) item).toLowerCase())
@@ -41,12 +48,17 @@ public class KnotxServerConfiguration {
 
     defaultFlow = new KnotxFlowConfiguration(config.getJsonObject("defaultFlow"));
     customFlow = new KnotxFlowConfiguration(config.getJsonObject("customFlow"));
-
+    fileUploadDirectory = config
+        .getString("fileUploadDirectory", BodyHandler.DEFAULT_UPLOADS_DIRECTORY);
     serverOptions = config.getJsonObject("serverOptions", new JsonObject());
   }
 
   public boolean displayExceptionDetails() {
     return displayExceptionDetails;
+  }
+
+  public JsonObject getCustomResponseHeader() {
+    return customResponseHeader;
   }
 
   public Set<String> getAllowedResponseHeaders() {
@@ -61,7 +73,12 @@ public class KnotxServerConfiguration {
     return customFlow;
   }
 
+  public String getFileUploadDirectory() {
+    return fileUploadDirectory;
+  }
+
   public JsonObject getServerOptions() {
     return serverOptions;
   }
 }
+