+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 All notable changes to Knot.x will be documented in this file.
 
 ## Unreleased
-
+List of changes that are finished but not yet released in any final version.
+ - [PR-345](https://github.com/Cognifide/knotx/pull/345) - Update to Vert.x 3.5 and RxJava 2
+ - [PR-320](https://github.com/Cognifide/knotx/pull/320) - Added KnotxServer configuration parameter to configure fileUploads folder
+ - [PR-347](https://github.com/Cognifide/knotx/pull/347) - Improved error logging on Server routing, repository connector & Http service adapters. Added custom headers configuration for requests/responses, etc.
+ - [PR-350](https://github.com/Cognifide/knotx/pull/350) - Cosmetic cleanup of default logback.xml config files. Fixed URLs to the bootstrap CSS on example app.
 
 ## Version 1.1.2
  - [PR-318](https://github.com/Cognifide/knotx/pull/318) - Knot.x returns exit code `30` in case of missing config
@@ -11,7 +15,6 @@ All notable changes to Knot.x will be documented in this file.
  - [PR-328](https://github.com/Cognifide/knotx/pull/328) - Knot.x ignore config parts related to not existing modules and allows to start the instance with warnings
  
 ## Version 1.1.1
-List of changes that are finished but not yet released in any final version.
  - [PR-316](https://github.com/Cognifide/knotx/pull/316) - Gateway Processor has access to request body
  - [PR-307](https://github.com/Cognifide/knotx/pull/307) - Fixed KnotxServer default configuration